requirements.txt
mmcv
argh==0.26.2
blessings==1.7
certifi
chardet==3.0.4
Click==7.0
configparser==4.0.2
docker-pycreds==0.4.0
gitdb2==2.0.6
GitPython==3.0.5
gql==0.2.0
graphql-core==1.1
idna==2.8
numpy
nltk==3.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
pathtools==0.1.2
Pillow==6.2.2
promise==2.3
psutil==5.6.7
python-dateutil==2.8.2
PyYAML==5.3
requests==2.22.0
sentry-sdk==0.14.0
shortuuid==0.5.0
six==1.16.0
smmap2==2.0.5
subprocess32==3.5.4
torch==1.12.1
torchvision==0.13.1
tqdm==4.41.1
urllib3==1.25.7
wandb==0.8.35
warmup-scheduler==0.3
watchdog==0.9.0
xxhash==1.4.3
